# [gql][performance-tuning][1/n] add `checkpoints_epoch_sequence_number` index to support graphql checkpoint queries