--- a/crates/sui-indexer/migrations/2023-08-19-044044_checkpoints/up.sql
+++ b/crates/sui-indexer/migrations/2023-08-19-044044_checkpoints/up.sql
@@ -24,6 +24,5 @@ CREATE TABLE checkpoints
     end_of_epoch_data                   bytea
 );
 
-CREATE INDEX checkpoints_epoch ON checkpoints (epoch);
-CREATE INDEX checkpoints_epoch_sequence_number ON checkpoints (epoch, sequence_number);
+CREATE INDEX checkpoints_epoch ON checkpoints (epoch, sequence_number);
 CREATE INDEX checkpoints_digest ON checkpoints USING HASH (checkpoint_digest);